More work on the dropped block. Added Lifespan component to destroy it over time and a spawnDroppedBlock command that creates the block with the correct mesh. Fixed player spawn location so the player doesn't spawn embedded in walls.

+++ src/main/java/org/terasology/logic/manager/PortalManager.java
@@ -114,7 +114,7 @@ public class PortalManager {
         return _portalStore.size() >= 1;
     }
 
-    public void addPortal(Vector3d loc) {
+    public void addPortal(Vector3f loc) {
         _portalStore.add(new Portal(loc));
     }
